+++ WeatherApp/Models/WeatherInfo/WeatherResponce.cs
@@ -11,19 +11,19 @@ namespace WeatherApp.Models
         /// Температура (°C).
         /// </summary>
         [JsonProperty(PropertyName = "temp")]
-        public decimal Temp { get; set; }
+        public double Temp { get; set; }
 
         /// <summary>
         /// Ощущаемая температура (°C).	
         /// </summary>
         [JsonProperty(PropertyName = "feels_like")]
-        public decimal FeelsLike { get; set; }
+        public double FeelsLike { get; set; }
 
         /// <summary>
         /// Температура воды (°C).Если в городе нет водоема, вернуть null	
         /// </summary>
         [JsonProperty(PropertyName = "temp_water")]
-        public decimal? TempWater { get; set; }
+        public double? TempWater { get; set; }
 
         /// <summary>
         /// Код расшифровки погодного описания
@@ -35,13 +35,13 @@ namespace WeatherApp.Models
         /// Скорость ветра (в м/с).	
         /// </summary>
         [JsonProperty(PropertyName = "wind_speed")]
-        public decimal WindSpeed { get; set; }
+        public double WindSpeed { get; set; }
 
         /// <summary>
         /// Скорость порывов ветра (в м/с).	
         /// </summary>
         [JsonProperty(PropertyName = "wind_gust")]
-        public decimal WindGust { get; set; }
+        public double WindGust { get; set; }
 
         /// <summary>
         /// Направление ветра
@@ -53,7 +53,7 @@ namespace WeatherApp.Models
         /// Давление (в мм рт. ст.).	
         /// </summary>
         [JsonProperty(PropertyName = "pressure_mm")]
-        public decimal PressureMm { get; set; }
+        public double PressureMm { get; set; }
 
         /// <summary>
         /// Тип осадков
